+++ src/common/api.ts
@@ -35,6 +35,7 @@ import {
   FileEdit,
   ClearProjectData,
   ProjectStartedData,
+  CloudflareTunnelStatus,
 } from '@common/types';
 
 export interface ApplicationAPI {
@@ -47,6 +48,9 @@ export interface ApplicationAPI {
   isManageServerSupported: () => boolean;
   startServer: (username?: string, password?: string) => Promise<boolean>;
   stopServer: () => Promise<boolean>;
+  startCloudflareTunnel: () => Promise<boolean>;
+  stopCloudflareTunnel: () => Promise<void>;
+  getCloudflareTunnelStatus: () => Promise<CloudflareTunnelStatus>;
   startProject: (baseDir: string) => void;
   stopProject: (baseDir: string) => void;
   restartProject: (baseDir: string, startupMode?: StartupMode) => void;