+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - extended context memory mode for Subagents to Full context or Last message
 - added server settings to control exposing the API and browser access
+- added basic authentication for the server access
+- added possibility to start Cloudflare tunnel to access the server
 
 ## [0.29.0]